+++ pkg/utils/test_key.go
@@ -2,10 +2,10 @@ package utils
 
 import "fmt"
 
-func GenTestKey(i uint16) []byte {
-	return []byte(fmt.Sprintf("key-%08d", i))
+func GenTestKey(i uint64) []byte {
+	return []byte(fmt.Sprintf("key-%010d", i))
 }
 
-func GenTestValue(i uint16) []byte {
-	return []byte(fmt.Sprintf("value-%08d", i))
+func GenTestValue(i uint64) []byte {
+	return []byte(fmt.Sprintf("value-%010d", i))
 }
